# perf: optimize dev server startup performance